requirements.txt
gitpython>=3.1.31
python-dotenv>=1.0.0
langchain>=0.0.350
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-ollama>=0.1.0
gradio>=3.37.0
faiss-cpu>=1.7.3
